hv: ipi: x86: update send_dest_ipi_mask() prototype
Align the prototype of send_dest_ipi_mask() on x86 with the RISC-V
definition. dest_mask is updated from uint32_t to uint64_t:

    From: void send_dest_ipi_mask(uint32_t dest_mask, uint32_t vector)
    To:   void send_dest_ipi_mask(uint64_t dest_mask, uint32_t vector)

On RISC-V, send_dest_ipi_mask() is implemented using SBI interfaces,
where the dest_mask is defined as "unsigned long" in the SBI spec.

Tracked-On: #8786
Acked-by: Wang, Yu1 <yu1.wang@intel.com>

+++ hypervisor/include/arch/x86/asm/lapic.h
@@ -102,7 +102,7 @@ void send_startup_ipi(uint16_t dest_pcpu_id, uint64_t cpu_startup_start_address)
  * @param[in]	dest_mask The mask of destination physical cpus
  * @param[in]	vector The vector of interrupt
  */
-void send_dest_ipi_mask(uint32_t dest_mask, uint32_t vector);
+void send_dest_ipi_mask(uint64_t dest_mask, uint32_t vector);
 
 /**
  * @brief Send an IPI to a single pCPU
